requirements.txt
redis[async]==5.1.1
openai==1.51.0
fastapi==0.115.0
uvicorn==0.31.0
pydantic==2.9.2
python-dotenv==1.0.1